DSi: fix printf format-string mismatches found by real compilation
Installed a generic arm-none-eabi-g++ (from apt -- the official
BlocksDS/Wonderful Toolchain download servers are blocked by this
sandbox's network policy) and compiled these files with -c against
the real blocksds/libnds headers cloned from GitHub, to check the
last two commits' code against an actual compiler instead of docs and
careful reading alone.

u32 resolves to "unsigned long" on this arm-eabi target, not
"unsigned int", so passing it to a bare %u was a real -Wformat
warning (and on some libc printf implementations, real garbage
output), not just a style nit. Cast to (unsigned) at the call site
instead of changing the format strings, matching the existing
(unsigned)sizeof(...) pattern already used a few lines below in
DsiBringup.cpp and in WiiBringup.cpp.

No other warnings or errors came out of the -Wall -Wextra -Wformat
compile of DsiEarlyVideo.cpp/DsiEarlyMemory.cpp/DsiBringup.cpp. This
is a syntax/type check only (no picolibc, no crt0/linkerscript, so no
link or .nds output) -- not a substitute for a real BlocksDS build,
but it does rule out the class of bug most likely from writing this
against docs rather than a compiler.

+++ src/dsi/tools/DsiBringup.cpp
@@ -53,8 +53,8 @@ void reportDsiMode()
 // 16 MB the console actually has -- that's the point of printing it here.
 void reportMemory()
 {
-	const u32 ceilingKb = dsiGetHeapCeiling() / 1024u;
-	const u32 committedKb = dsiGetHeapCommitted() / 1024u;
+	const unsigned ceilingKb = (unsigned)(dsiGetHeapCeiling() / 1024u);
+	const unsigned committedKb = (unsigned)(dsiGetHeapCommitted() / 1024u);
 	std::printf("HEAP    %u KB ceiling (enforced), %u KB committed\n", ceilingKb, committedKb);
 	std::printf("BUDGET  8192 KB target; PS2 used ~14336 of 32768\n");
 }
